fix(python): ensure consistent handling of 1D numpy arrays with respect to other sequences
